swapped stock data API to FinnHub
yFinance was unstable and sluggish - stopped working. Moved to something more stable with a proper API and my own personal key. created .env file for backend which houses the key, wont be put up to GitHub for security reasons. and added the key to render enviornmental variables on renders dashboard for my webiste